# docs/api/gateway.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PHS — API pública (gateway)",
    "version": "0.1.0",
    "description": "Contrato que usa la aplicación web. Toda ruta entra por el gateway, que valida la sesión y reenvía a los servicios."
  },
  "paths": {
...
    "/api/v1/catalog/service-types": {
      "get": {
        "summary": "Catálogo de tipos de servicio",
        "tags": [
          "Catálogos"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Tipos activos e inactivos.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "code": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "active": {
                        "type": "boolean"
                      }
                    },
                    "required": [
                      "code",
                      "name",
                      "active"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión válida (`authentication_required`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso (`forbidden`), contraseña temporal sin cambiar (`password_change_required`) u origen no permitido (`origin_not_allowed`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "502": {
            "description": "El servicio de identidad no respondió (`identity_unavailable`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Agregar tipo de servicio",
        "tags": [
          "Catálogos"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "description": "Requiere un usuario administrador.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "pattern": "^[a-z][a-z0-9_]{1,39}$"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  }
                },
                "required": [
                  "code",
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tipo creado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "active": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "code",
                    "name",
                    "active"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos (`invalid_request`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión válida (`authentication_required`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso (`forbidden`), contraseña temporal sin cambiar (`password_change_required`) u origen no permitido (`origin_not_allowed`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Código o nombre ya usados (`code_taken`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "502": {
            "description": "El servicio de identidad no respondió (`identity_unavailable`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/service-types/{code}": {
      "patch": {
        "summary": "Activar o desactivar un tipo de servicio",
        "tags": [
          "Catálogos"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "description": "Requiere un usuario administrador.",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "active": {
                    "type": "boolean",
                    "description": "false lo oculta para nuevos proyectos; los existentes lo conservan."
                  }
                },
                "required": [
                  "active"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tipo actualizado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "active": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "code",
                    "name",
                    "active"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos (`invalid_request`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión válida (`authentication_required`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso (`forbidden`), contraseña temporal sin cambiar (`password_change_required`) u origen no permitido (`origin_not_allowed`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "No existe o está fuera del alcance del usuario (`not_found`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "502": {
            "description": "El servicio de identidad no respondió (`identity_unavailable`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "phs_session",
        "description": "Cookie de sesión emitida por POST /api/v1/session."
      },
      "gatewaySignature": {
        "type": "apiKey",
        "in": "header",
        "name": "x-phs-internal-auth",
        "description": "Identidad firmada por el gateway (HMAC-SHA256, válida 60 s). No se acepta desde el navegador."
      }
    }
  }
}